Update box.py
The current implementation of the Box class has a potential issue related to the sides attribute. When setting a side (top, bottom, left, right) that has already been set, the sides attribute is incremented again, which is not desired behavior. This can lead to incorrect counts of the sides and incorrect increment of BOXES_DONE